Bug 1946361 - Add debug prefs for missing snapshots. r=gw
Differential Revision: https://phabricator.services.mozilla.com/D237008

+++ gfx/config/gfxVars.h
@@ -51,7 +51,7 @@ class gfxVarReceiver;
   _(WebglAllowCoreProfile, bool, true)                             \
   _(WebglAllowWindowsNativeGl, bool, false)                        \
   _(WebRenderMaxPartialPresentRects, int32_t, 0)                   \
-  _(WebRenderDebugFlags, int32_t, 0)                               \
+  _(WebRenderDebugFlags, uint64_t, 0)                              \
   _(WebRenderBoolParameters, int32_t, 0)                           \
   _(WebRenderBatchingLookback, int32_t, 10)                        \
   _(WebRenderBlobTileSize, int32_t, 256)                           \
